refactor(super-admin): deduplicate fair-use edit helpers (AYC-71)
Applied via @cursor push command
